Seperate paths from Contours.py
